requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
python-dotenv==1.0.1
dj-rest-auth==5.0.2
django-allauth==0.61.1